DAOTHER-8059 fix cron job, rename default lock
