+++ deploy/example-crontab
@@ -2,10 +2,11 @@
 # (may need 6th parameter "root" before python3)
 # runs every 4 hours
 
-# remove sessions
+# remove sessions (and locks over 1 hour old)
 # with no parameter, removes only unexpired sessions (as set in config.yml kv_max_seconds)
 # test with * * * * * and parameter 0 to remove all sessions
 0 0/4 * * * python3 -m rcsb.app.file.Sessions
 
 # remove locks over 1 hour old
-0 0/4 * * * python3 -m rcsb.app.file.TernaryLock True 3600
+# (moved to Sessions)
+# 0 0/4 * * * python3 -m rcsb.app.file.RedisLock True 3600
