+++ rcsb/app/config/config.yml
@@ -14,13 +14,14 @@ configuration:
   SESSION_DIR_PATH: rcsb/app/tests-file/data/sessions
   SHARED_LOCK_PATH: rcsb/app/tests-file/data/shared-locks  # default or ternary lock only
   LOCK_TRANSACTIONS: True
-  LOCK_TYPE: redis  # default, ternary, or redis (need not agree with kv mode)
+  LOCK_TYPE: redis # soft, ternary, or redis (need not agree with kv mode)
+  REMOVE_LOCKS_AFTER: 3600 # seconds ... remove after 1 hour of inactivity
   # database parameters
   KV_MODE: sqlite # redis or sqlite, redis for multiple machines or containers, sqlite possible for one machine or container only
   REDIS_HOST: localhost # localhost, redis, or url
   KV_SESSION_TABLE_NAME: sessions
   KV_MAP_TABLE_NAME: map
-  KV_LOCK_TABLE_NAME: lock
+  KV_LOCK_TABLE_NAME: lock # redis lock only
   KV_MAX_SECONDS: 14400 # session duration
   KV_FILE_PATH: ./kv.sqlite # sqlite only
   # file parameters
