+++ rcsb/app/file/IoUtility.py
@@ -34,7 +34,7 @@ if locktype == "redis":
 elif locktype == "ternary":
     from rcsb.app.file.TernaryLock import Locking
 else:
-    from rcsb.app.file.DefaultLock import Locking
+    from rcsb.app.file.SoftLock import Locking
 
 
 logging.basicConfig(
