+++ rcsb/app/file/main.py
@@ -30,7 +30,7 @@ if locktype == "redis":
 elif locktype == "ternary":
     from rcsb.app.file.TernaryLock import Locking
 else:
-    from rcsb.app.file.DefaultLock import Locking
+    from rcsb.app.file.SoftLock import Locking
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
