+++ tox.ini
@@ -27,9 +27,9 @@ commands =
     pylint --disable=R,C --reports=n --rcfile={toxinidir}/pylintrc  rcsb/app/client setup.py
 ;    black --diff --line-length 185  rcsb/app/file setup.py
     coverage erase
-    coverage run --parallel-mode --omit="*__init__.py,*client.py,*gui.py,*DefaultLock.py" --source="rcsb/app/client" -m unittest discover -v --start-directory rcsb/app/tests-client --pattern "test*.py"
-    coverage run --parallel-mode --omit="*__init__.py,*client.py,*gui.py,*DefaultLock.py" --source="rcsb/app/file" -m unittest discover -v --start-directory rcsb/app/tests-client --pattern "test*.py"
-    coverage run --parallel-mode --omit="*__init__.py,*client.py,*gui.py,*DefaultLock.py" --source="rcsb/app/file" -m unittest discover -v --start-directory rcsb/app/tests-file --pattern "test*.py"
+    coverage run --parallel-mode --omit="*__init__.py,*client.py,*gui.py,*Lock.py" --source="rcsb/app/client" -m unittest discover -v --start-directory rcsb/app/tests-client --pattern "test*.py"
+    coverage run --parallel-mode --omit="*__init__.py,*client.py,*gui.py,*Lock.py" --source="rcsb/app/file" -m unittest discover -v --start-directory rcsb/app/tests-client --pattern "test*.py"
+    coverage run --parallel-mode --omit="*__init__.py,*client.py,*gui.py,*Lock.py" --source="rcsb/app/file" -m unittest discover -v --start-directory rcsb/app/tests-file --pattern "test*.py"
     echo " ------- Consolidating {envname} data ----------"
     coverage combine
     echo " ------- Building {envname} reports ----------"
